requirements.txt
paramiko
scp
requests[security]
pexpect
fastapi
fastapi-mail
fastapi-login
uvicorn
jinja2
aiofiles
zipfile36
uncompyle6
autopep8
openpyxl
configparser
coverage
Markdown
python-dotenv
python-multipart
pytest
pytest-cov
pytest-html
pytest-mock
pytest-order
python-docx
python-gitlab
redis
pymongo
mysqlclient
ldap3==2.5.0